--- apps/blog/src/app/_components/Blog/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Blog";
