--- apps/blog/src/app/_components/Extra/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Extra";
