--- apps/blog/src/app/_components/Introduce/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Introduce";
