--- apps/blog/src/app/_components/Maintenance/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Maintenance";
