--- apps/blog/src/app/_components/Projects/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Projects";
