+++ apps/blog/src/widgets/home-blog/index.ts
@@ -1,0 +1,1 @@
+export { default as Blog } from "./ui/Blog";
