+++ apps/blog/src/widgets/home-extra/index.ts
@@ -1,0 +1,1 @@
+export { default as Extra } from "./ui/Extra";
