+++ apps/blog/src/widgets/home-introduce/index.ts
@@ -1,0 +1,1 @@
+export { default as Introduce } from "./ui/Introduce";
