+++ apps/blog/src/widgets/home-projects/index.ts
@@ -1,0 +1,1 @@
+export { default as Projects } from "./ui/Projects";
