+++ apps/blog/src/widgets/maintenance/index.ts
@@ -1,0 +1,1 @@
+export { default as Maintenance } from "./ui/Maintenance";
